refactor: AggregationScreenの左パネルをSettingsPanelに分離
DataSummary, CrossConfig, WeightInfo等の設定UIを
SettingsPanel.tsxに切り出し、AggregationScreenの責務を軽減。
